Warn about config expansions as we do for other expansions.
If an expanded value overrides an explicit value, users who do not know the contents of the expansion may be surprised. We already warned about this for hard-coded expansions, and this is now applicable for --config expansions as well.

This will only warn when a single-valued option has its value replaced. Options that accumulate multiple values in a list (e.g., --copt) will silently include both explicit and expanded values.

RELNOTES: None.
PiperOrigin-RevId: 179857526
GitOrigin-RevId: 0421d7d8566a6fbe35e17a1edc3ab4d622aa6c9e
Change-Id: Ie028995d2c4cbb90614ea8094b662d1b6e319241