Revert "Fix annotation processors working w/ generated code on OpenJDK 9 toolchain"
This change is no longer needed, following http://r.android.com/943770
in April 2019, which upgraded external/dagger2 from a 2015 version of
auto-common to version 0.10
(https://github.com/google/auto/releases/tag/auto-common-0.10). This
auto-common version does not output code using
javax.annotation.Generated if it is is not available, so the module
that provides (java.xml.ws.annotation) it can be safely removed.

This is beneficial since the module is deprecated in OpenJDK 9 and is
removed in later versions.

N.B. The change being reverted was to Android.mk, this removes the
corresponding lines from the migrated Android.bp.

This reverts commit 66b92c8c9172780c5794f47cbe7ede4eeee009b7.

Bug: 62050818
Bug: 131683177
Test: EXPERIMENTAL_JAVA_LANGUAGE_LEVEL_9=true make droid
Change-Id: Id964a685dbadb21f414b15e029d5e02a40ece83f